feat(TableRow): add staggerIndex prop for opt-in entrance animation
Promote the previously internal rowEntering/staggerIndex helpers into a
public API on TableRow. Pass staggerIndex={i} to animate a row's cells
with a staggered slide-up-and-fade entrance; omit to disable.

- New file: lib/components/Table/TableRow.css.ts now owns the animation
  keyframe, rowEntering class, and staggerIndex CSS variable. Moved out
  of DataTable.css.ts so TableRow (a primitive) does not depend on
  DataTable.
- TableRow merges rowEntering className and assignInlineVars style when
  staggerIndex is a number, using clsx for className composition.
- DataTable stories now use the prop instead of importing the CSS
  primitives directly.
- Changeset updated to document the new prop.
