+++ lib/components/DataTable/DataTable.css.ts
@@ -1,4 +1,4 @@
-import { createVar, globalStyle, keyframes, style } from '@vanilla-extract/css';
+import { globalStyle, keyframes, style } from '@vanilla-extract/css';
 
 import { overdriveTokens as vars } from '../../themes/theme.css';
 
@@ -44,33 +44,4 @@ globalStyle(`${scrollContainer} [role="columnheader"]`, {
 globalStyle(`${innerWrapper} [role="gridcell"]::before`, {
 	left: 0,
 	right: 0,
-});
-
-/*
- * Row entrance animation — staggered slide-up with fade.
- *
- * Apply `rowEntering` as className on a <TableRow> and set the
- * `staggerIndex` CSS variable to the row's index via assignInlineVars.
- * The animation targets child gridcells since TableRow uses
- * display:contents (no box to animate directly).
- */
-
-export const staggerIndex = createVar();
-
-const slideUp = keyframes({
-	from: {
-		opacity: 0,
-		transform: 'translateY(12px)',
-	},
-	to: {
-		opacity: 1,
-		transform: 'translateY(0)',
-	},
-});
-
-export const rowEntering = style({});
-
-globalStyle(`${rowEntering} > [role="gridcell"]`, {
-	animation: `${slideUp} 300ms ${vars.animation.easing.decelerate} both`,
-	animationDelay: `calc(${staggerIndex} * 50ms)`,
 });
